Star At Abim

+++ src/commands/Sıfırlama/cezapuansil.js
@@ -34,7 +34,7 @@ return;
 }
 message.react(green)
 message.lineReply(`${green} ${member} üyesinin ceza puanı ${message.author} tarafından \`${moment(Date.now()).format("LLL")}\` tarihinde temizlendi!`)
-await cezapuan.deleteMany({})
+await cezapuan.deleteMany({userID: member.user.id, guildID: message.guild.id})
 await ceza.deleteMany({userID: member.user.id, guildID: message.guild.id})
 }
 };
